fastapi/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
decorator==4.4.2
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.4
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.10.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
Jinja2==3.1.4
jiter==0.7.0
MarkupSafe==3.0.2
moviepy==1.0.3
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numpy<2
openai==0.28.0
packaging==24.2
pillow==11.0.0
proglog==0.1.10
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
sniffio==1.3.1
starlette==0.41.2
sympy==1.13.3
tokenizers==0.20.3
torch==2.2.2
tqdm==4.67.0
transformers==4.46.2
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
yarl==1.17.1
